+++ src/main/java/com/github/minecraftschurlimods/arsmagicalegacy/common/block/altar/AltarCoreBlockEntity.java
@@ -285,6 +285,7 @@ public class AltarCoreBlockEntity extends BlockEntity implements IEtheriumConsum
                     .ifRight(ArsMagicaLegacy.LOGGER::warn)
                     .left()
                     .orElse(null);
+            requestModelDataUpdate();
         }
         super.load(pTag);
     }