+++ src/main/java/com/github/minecraftschurlimods/arsmagicalegacy/compat/patchouli/StairBlockStateMatcher.java
@@ -4,7 +4,6 @@ import com.github.minecraftschurlimods.arsmagicalegacy.api.altar.AltarStructureM
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.Registry;
-import net.minecraft.core.RegistryAccess;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.StairBlock;
@@ -21,12 +20,12 @@ class StairBlockStateMatcher implements IStateMatcher {
     public StairBlockStateMatcher(Direction direction, Half half) {
         this.direction = direction;
         this.half = half;
-        this.predicate = (blockGetter, blockPos, state) -> RegistryAccess.BUILTIN.get().registryOrThrow(AltarStructureMaterial.REGISTRY_KEY).stream().anyMatch(mat -> state.is(mat.stair())) && state.getValue(StairBlock.FACING) == direction && state.getValue(StairBlock.HALF) == half;
+        this.predicate = (blockGetter, blockPos, state) -> PatchouliCompat.getRegistry(blockGetter, AltarStructureMaterial.REGISTRY_KEY).stream().anyMatch(mat -> state.is(mat.stair())) && state.getValue(StairBlock.FACING) == direction && state.getValue(StairBlock.HALF) == half;
     }
 
     @Override
     public BlockState getDisplayedState(long ticks) {
-        Registry<AltarStructureMaterial> registry = RegistryAccess.BUILTIN.get().registryOrThrow(AltarStructureMaterial.REGISTRY_KEY);
+        Registry<AltarStructureMaterial> registry = PatchouliCompat.getRegistry(null, AltarStructureMaterial.REGISTRY_KEY);
         AltarStructureMaterial mat = registry.stream().toArray(AltarStructureMaterial[]::new)[Math.toIntExact(ticks / 20) % registry.size()];
         if (mat == null) return Blocks.AIR.defaultBlockState();
         return mat.stair().defaultBlockState().setValue(StairBlock.FACING, direction).setValue(StairBlock.HALF, half);