+++ changelog.md
@@ -1,5 +1,7 @@
-# Version 1.19.2-1.1.5
+# Version 1.19.2-1.1.6
 
 ## Fixes
 
-- Fixed crash.
+- Fixed divide by 0 in patchouli book.
+- Fixed altar core not updating camo state correctly.
+- Fixed `KnowledgeHolder#skillPoints` being immutable.